worked on resolving comments
